chore: rename browser to open
