+++ packages/astro/src/cli/open.ts
